Use uidToEntity all over net.cpp
This should improve performance significantly (#302).

Also remove the getEntityWithUID method from map. It was only used in
one location and redundant with uidToEntity.

The use of uidToEntity in net.cpp is currently unchecked (i.e. will fail
segfaultily if no entity with the given uid is found). Cursory testing
did not exhibit such a crash, but it should probably be checked
properly anyway.